cleanup(core): move the string-width helpers from link.rs to tui utils
display_width and fit_with_ellipsis are generic string utilities, not
Link-widget concerns — review feedback asked for them to live in a
shared home. Move them to tui/utils.rs alongside the other formatting
helpers and update the callers.
